+++ src/pop/integrations/crewai/__init__.py
@@ -1,0 +1,9 @@
+from pop.integrations.crewai.bindings import (
+    bind_crewai_agent_kwargs,
+    create_crewai_agent_from_persona,
+)
+
+__all__ = [
+    "bind_crewai_agent_kwargs",
+    "create_crewai_agent_from_persona",
+]